+++ packages/tanstack-react-query/src/hooks/useQuery.ts
@@ -1,8 +1,7 @@
-import { parseQuery, type CompilableQuery } from '@powersync/common';
+import { type CompilableQuery } from '@powersync/common';
 import { usePowerSync } from '@powersync/react';
-import React from 'react';
-
 import * as Tanstack from '@tanstack/react-query';
+import { usePowerSyncQueries } from './usePowerSyncQueries.js';
 
 export type PowerSyncQueryOptions<T> = {
   query?: string | CompilableQuery<T>;
@@ -11,12 +10,44 @@ export type PowerSyncQueryOptions<T> = {
 
 export type UseBaseQueryOptions<TQueryOptions> = TQueryOptions & PowerSyncQueryOptions<any>;
 
+/**
+ *
+ * Uses the `queryFn` to execute the query. No different from the base `useQuery` hook.
+ *
+ * @example
+ * ```
+ * const { data, error, isLoading } = useQuery({
+ *   queryKey: ['lists'],
+ *   queryFn: getTodos,
+ * });
+ * ```
+ */
 export function useQuery<TData = unknown, TError = Tanstack.DefaultError>(
   options: UseBaseQueryOptions<Tanstack.UseQueryOptions<TData, TError>> & { query?: undefined },
   queryClient?: Tanstack.QueryClient
 ): Tanstack.UseQueryResult<TData, TError>;
 
-// Overload when 'query' is present
+/**
+ *
+ * Uses the `query` to execute the PowerSync query.
+ *
+ * @example
+ * ```
+ * const { data, error, isLoading } = useQuery({
+ *   queryKey: ['lists'],
+ *   query: 'SELECT * from lists where id = ?',
+ *   parameters: ['id-1']
+ * });
+ * ```
+ *
+ * @example
+ * ```
+ * const { data, error, isLoading } = useQuery({
+ *   queryKey: ['lists'],
+ *   query: compilableQuery,
+ * });
+ * ```
+ */
 export function useQuery<TData = unknown, TError = Tanstack.DefaultError>(
   options: UseBaseQueryOptions<Tanstack.UseQueryOptions<TData[], TError>> & { query: string | CompilableQuery<TData> },
   queryClient?: Tanstack.QueryClient
@@ -29,12 +60,36 @@ export function useQuery<TData = unknown, TError = Tanstack.DefaultError>(
   return useQueryCore(options, queryClient, Tanstack.useQuery);
 }
 
+/**
+ *
+ * Uses the `queryFn` to execute the query. No different from the base `useSuspenseQuery` hook.
+ *
+ * @example
+ * ```
+ * const { data } = useSuspenseQuery({
+ *   queryKey: ['lists'],
+ *   queryFn: getTodos,
+ * });
+ * ```
+ */
 export function useSuspenseQuery<TData = unknown, TError = Tanstack.DefaultError>(
   options: UseBaseQueryOptions<Tanstack.UseSuspenseQueryOptions<TData, TError>> & { query?: undefined },
   queryClient?: Tanstack.QueryClient
 ): Tanstack.UseSuspenseQueryResult<TData, TError>;
 
-// Overload when 'query' is present
+/***
+ *
+ * Uses the `query` to execute the PowerSync query.
+ *
+ * @example
+ * ```
+ * const { data } = useSuspenseQuery({
+ *   queryKey: ['lists'],
+ *   query: 'SELECT * from lists where id = ?',
+ *   parameters: ['id-1']
+ * });
+ * ```
+ */
 export function useSuspenseQuery<TData = unknown, TError = Tanstack.DefaultError>(
   options: UseBaseQueryOptions<Tanstack.UseSuspenseQueryOptions<TData[], TError>> & {
     query: string | CompilableQuery<TData>;
@@ -65,92 +120,23 @@ function useQueryCore<
     throw new Error('PowerSync is not available');
   }
 
-  let error: Error | undefined = undefined;
+  const { query, parameters, queryKey, ...resolvedOptions } = options;
 
-  const [tables, setTables] = React.useState<string[]>([]);
-  const { query, parameters = [], ...resolvedOptions } = options;
-
-  let sqlStatement = '';
-  let queryParameters = [];
-
-  if (query) {
-    try {
-      const parsedQuery = parseQuery(query, parameters);
-
-      sqlStatement = parsedQuery.sqlStatement;
-      queryParameters = parsedQuery.parameters;
-    } catch (e) {
-      error = e;
-    }
-  }
-
-  const stringifiedParams = JSON.stringify(queryParameters);
-  const stringifiedKey = JSON.stringify(options.queryKey);
-
-  const fetchTables = async () => {
-    try {
-      const tables = await powerSync.resolveTables(sqlStatement, queryParameters);
-      setTables(tables);
-    } catch (e) {
-      error = e;
-    }
-  };
-
-  React.useEffect(() => {
-    if (error || !query) return () => {};
-
-    (async () => {
-      await fetchTables();
-    })();
-
-    const l = powerSync.registerListener({
-      schemaChanged: async () => {
-        await fetchTables();
-        queryClient.invalidateQueries({ queryKey: options.queryKey });
-      }
-    });
-
-    return () => l?.();
-  }, [powerSync, sqlStatement, stringifiedParams]);
-
-  const queryFn = React.useCallback(async () => {
-    if (error) {
-      return Promise.reject(error);
-    }
-
-    try {
-      return typeof query == 'string' ? powerSync.getAll<TData>(sqlStatement, queryParameters) : query.execute();
-    } catch (e) {
-      return Promise.reject(e);
-    }
-  }, [powerSync, query, parameters, stringifiedKey]);
-
-  React.useEffect(() => {
-    if (error || !query) return () => {};
-
-    const abort = new AbortController();
-    powerSync.onChangeWithCallback(
+  const [{ queryFn }] = usePowerSyncQueries(
+    [
       {
-        onChange: () => {
-          queryClient.invalidateQueries({
-            queryKey: options.queryKey
-          });
-        },
-        onError: (e) => {
-          error = e;
-        }
-      },
-      {
-        tables,
-        signal: abort.signal
+        query,
+        parameters,
+        queryKey
       }
-    );
-    return () => abort.abort();
-  }, [powerSync, queryClient, stringifiedKey, tables]);
+    ],
+    queryClient
+  );
 
   return useQueryFn(
     {
       ...(resolvedOptions as TQueryOptions),
+      queryKey,
       queryFn: query ? queryFn : resolvedOptions.queryFn
     } as TQueryOptions,
     queryClient