+++ packages/capacitor/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/capacitor
 
+## 0.1.3
+
+### Patch Changes
+
+- 4c66487: Fixed readTransaction method throwing "not allowed in read-only mode" errors
+  - @powersync/web@1.28.2
+
+## 0.1.2
+
+### Patch Changes
+
+- 3e4a25c: Don't minify releases, enable source maps.
+- Updated dependencies [3e4a25c]
+  - @powersync/web@1.28.1
+
 ## 0.1.1
 
 ### Patch Changes