+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @powersync/react
 
+## 1.8.2
+
+### Patch Changes
+
+- 3e4a25c: Don't minify releases, enable source maps.
+- Updated dependencies [3e4a25c]
+  - @powersync/common@1.41.1
+
 ## 1.8.1
 
 ### Patch Changes